+++ Classes/ViewHelpers/SlugViewHelper.php
@@ -5,7 +5,6 @@ use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Fluid\Core\Rendering\RenderingContextInterface;
 use TYPO3\CMS\Fluid\Core\ViewHelper\AbstractViewHelper;
 use TYPO3\CMS\Fluid\Core\ViewHelper\Facets\CompilableInterface;
-use TYPO3\CMS\Frontend\ContentObject\ContentObjectRenderer;
 
 use Cocur\Slugify\Slugify;
 
@@ -20,10 +19,9 @@ class SlugViewHelper extends AbstractViewHelper implements CompilableInterface {
    *
    * @param string $value The value to generate a slug from
    * @param string $separator The optional separator, defaults to `-`
-   *
    * @return string The rendered output
    */
-  public function render($value, $separator = '-') {
+  public function render(string $value, string $separator = '-') {
     return static::renderStatic(
       [
         'value'     => $value,
@@ -35,10 +33,11 @@ class SlugViewHelper extends AbstractViewHelper implements CompilableInterface {
   }
 
   /**
+   * The view helper render static function.
+   *
    * @param array $arguments The arguments
    * @param callable $renderChildrenClosure The render children closure
    * @param RenderingContextInterface $renderingContext The rendering context
-   *
    * @return string The rendered output
    */
   public static function renderStatic(array $arguments, \Closure $renderChildrenClosure, RenderingContextInterface $renderingContext) {
@@ -59,13 +58,9 @@ class SlugViewHelper extends AbstractViewHelper implements CompilableInterface {
    *
    * @param string $value The value to generate a slug from
    * @param string $separator The optional separator, defaults to `-`
-   *
    * @return string The slug
    */
-  protected static function createSlug($value, $separator = '-') {
-    $value     = (string) $value;
-    $separator = (string) $separator;
-
+  protected static function createSlug(string $value, string $separator = '-') {
     $slugify = GeneralUtility::makeInstance(Slugify::class);
     $slug    = $slugify->slugify($value, $separator);