+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Notable changes will be documented in this file. The project adheres to [Semantic Versioning].
 
+* Cleaned up `SlugViewHelper`
+
 6.3.1
 -----
 
@@ -91,7 +93,7 @@ Notable changes will be documented in this file. The project adheres to [Semanti
 3.0.0
 -----
 
-* Added `Slug` View Helper
+* Added `SlugViewHelper`
 * Added `Cocur/Slugify` as dependency
 * Refactored TypoScript
 * Updated constants and configuration